requirements.txt
Flask~=2.0.1
google-cloud-bigquery~=2.13.1
requests~=2.23.0
